requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
spotipy>=2.23.0
jupyter>=1.0.0
joblib>=1.3.0
streamlit>=1.32.0 